+++ src/main/java/com/github/nibiroo/GMotors/repository/AnnouncementRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
@@ -17,8 +18,8 @@ public interface AnnouncementRepository extends JpaRepository<Announcement, Long
     List<Announcement> findAllThanZero();
 
     @Query(value = "SELECT * FROM announcement WHERE KM >= ?1 AND KM <= ?2", nativeQuery = true)
-    List<Announcement> findAllThanKmInRangeKm(Float smallerKm, Float biggerKm);
+    List<Announcement> findAllThanKmInRangeKm(BigDecimal smallerKm, BigDecimal biggerKm);
 
     @Query(value = "SELECT * FROM announcement WHERE PRICE >= ?1 AND PRICE <= ?2", nativeQuery = true)
-    List<Announcement> findAllThanPriceInRangePrice(Float smallerPrice, Float biggerPrice);
+    List<Announcement> findAllThanPriceInRangePrice(BigDecimal smallerPrice, BigDecimal biggerPrice);
 }
