+++ src/main/java/com/github/nibiroo/GMotors/service/AnnouncementService.java
@@ -3,7 +3,7 @@ package com.github.nibiroo.GMotors.service;
 import com.github.nibiroo.GMotors.entity.Announcement;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -16,9 +16,9 @@ public interface AnnouncementService {
 
     List<Announcement> findAllAnnouncementsWithVehiclesNews();
 
-    List<Announcement> findAllAnnouncementsWithVehiclesWithRangeKm(Float smallerKm, Float biggerKm);
+    List<Announcement> findAllAnnouncementsWithVehiclesWithRangeKm(BigDecimal smallerKm, BigDecimal biggerKm);
 
-    List<Announcement> findAllAnnouncementsWithVehiclesWithRangePrice(Float smallerPrice, Float biggerPrice);
+    List<Announcement> findAllAnnouncementsWithVehiclesWithRangePrice(BigDecimal smallerPrice, BigDecimal biggerPrice);
 
     Announcement save(Announcement announcement);
