+++ src/main/java/com/github/nibiroo/GMotors/service/impl/AnnouncementServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
@@ -42,12 +43,12 @@ public class AnnouncementServiceImpl implements AnnouncementService {
     }
 
     @Override
-    public List<Announcement> findAllAnnouncementsWithVehiclesWithRangeKm(Float smallerKm, Float biggerKm) {
+    public List<Announcement> findAllAnnouncementsWithVehiclesWithRangeKm(BigDecimal smallerKm, BigDecimal biggerKm) {
         return this.announcementRepository.findAllThanKmInRangeKm(smallerKm, biggerKm);
     }
 
     @Override
-    public List<Announcement> findAllAnnouncementsWithVehiclesWithRangePrice(Float smallerPrice, Float biggerPrice) {
+    public List<Announcement> findAllAnnouncementsWithVehiclesWithRangePrice(BigDecimal smallerPrice, BigDecimal biggerPrice) {
         return this.announcementRepository.findAllThanPriceInRangePrice(smallerPrice, biggerPrice);
     }
